refactor(draw): move ol-draw logic to draw composable
