+++ src/components/map/map-container.vue
@@ -14,7 +14,7 @@ import Map3dControl from '../map-controls/map-3d.vue'
 import FullscreenControl from '../map-controls/fullscreen-control.vue'
 import ZoomControl from '../map-controls/zoom-control.vue'
 import ZoomToExtentControl from '../map-controls/zoom-to-extent-control.vue'
-import useOlDraw from '@/composables/map/ol-draw.composable'
+import useDraw from '@/composables/draw/draw.composable'
 
 const appStore = useAppStore()
 const { embedded } = storeToRefs(appStore)
@@ -22,8 +22,8 @@ const map = useMap()
 const mapContainer = ref(null)
 const olMap = map.createMap()
 // add draw layer after map init to allow restoring draw features
-const olDraw = useOlDraw()
-olDraw.addDrawLayer(olMap)
+const draw = useDraw()
+draw.addDrawLayer(olMap)
 
 withDefaults(
   defineProps<{
